:bug: message queue now create a que to an  existing file with a unique id
